somewhat biggish container management PR

+++ common/db/migrations/066_remove_top_container_legacy_restricted.rb
@@ -3,7 +3,7 @@ require_relative 'utils'
 Sequel.migration do
   up do
     alter_table(:top_container) do
-      drop_column(:legacy_restricted)     
+      drop_column(:legacy_restricted) unless AppConfig[:plugins].include?("container_management") 
     end
   end
